+++ BL/Interfaces/DTO/IstbURLTypeDTO.cs
@@ -8,7 +8,7 @@ namespace BL.Interfaces
 {
     public interface IstbURLTypeDTO
     {
-        Int64 URLTypeId { get; set; }
+        int URLTypeId { get; set; }
         string Name { get; set; }
         string Description { get; set; }
     }
